# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Fintech Microservice API
  description: API specification for the fintech microservice application
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Auth Service
  - url: http://localhost:3002
    description: Accounts Service
  - url: http://localhost:3003
    description: Transfer Service
  - url: http://localhost:3004
    description: Ledger Service
components:
  schemas:
...
    TransferRequest:
      type: object
      properties:
        fromAccountId:
          type: integer
        toAccountId:
          type: integer
        amount:
          type: number
      required:
        - fromAccountId
        - toAccountId
        - amount
    TransferResponse:
      type: object
      properties:
        message:
          type: string
        transferId:
          type: integer
        fromTransactionId:
          type: integer
        toTransactionId:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /transfer:
    post:
      summary: Initiate fund transfer
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '200':
          description: Transfer already processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  transferId:
                    type: integer
                  status:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transfer/{id}/status:
    get:
      summary: Get transfer status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transfer status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transfer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health endpoints (available on all services)
...